refactor: move representative runner into workflow shell
