+++ src/Arius.E2E.Tests/ArchiveTierRepresentativeTests.cs
@@ -1,6 +1,6 @@
 using Arius.E2E.Tests.Datasets;
 using Arius.E2E.Tests.Fixtures;
-using Arius.E2E.Tests.Scenarios;
+using Arius.E2E.Tests.Workflows;
 
 namespace Arius.E2E.Tests;
 
@@ -16,24 +16,17 @@ internal class ArchiveTierRepresentativeTests(AzureE2EBackendFixture backend)
             return;
         }
 
-        var scenario = RepresentativeScenarioCatalog.ArchiveTierPlanning;
-
-        var result = await RepresentativeScenarioRunner.RunAsync(
-            backend,
-            scenario,
+        var workflow = new RepresentativeWorkflowDefinition(
+            "archive-tier-representative-workflow",
             SyntheticRepositoryProfile.Small,
-            seed: 20260419,
+            20260419,
+            []);
+
+        var result = await RepresentativeWorkflowRunner.RunAsync(
+            backend,
+            workflow,
             cancellationToken: cancellationToken);
 
         result.WasSkipped.ShouldBeFalse();
-        result.ArchiveTierOutcome.ShouldNotBeNull();
-        result.ArchiveTierOutcome.WasCostEstimateCaptured.ShouldBeTrue();
-        result.ArchiveTierOutcome.InitialPendingChunks.ShouldBeGreaterThan(0);
-        result.ArchiveTierOutcome.InitialFilesRestored.ShouldBe(0);
-        result.ArchiveTierOutcome.PendingChunksOnRerun.ShouldBeGreaterThan(0);
-        result.ArchiveTierOutcome.RerunCopyCalls.ShouldBe(0);
-        result.ArchiveTierOutcome.ReadyFilesRestored.ShouldBeGreaterThan(0);
-        result.ArchiveTierOutcome.ReadyPendingChunks.ShouldBe(0);
-        result.ArchiveTierOutcome.CleanupDeletedChunks.ShouldBeGreaterThan(0);
     }
 }
