+++ src/Arius.E2E.Tests/RepresentativeArchiveRestoreTests.cs
@@ -1,6 +1,6 @@
 using Arius.E2E.Tests.Datasets;
 using Arius.E2E.Tests.Fixtures;
-using Arius.E2E.Tests.Scenarios;
+using Arius.E2E.Tests.Workflows;
 
 namespace Arius.E2E.Tests;
 
@@ -8,9 +8,9 @@ internal class RepresentativeArchiveRestoreTests
 {
     [Test]
     [CombinedDataSources]
-    public async Task Representative_Scenario_Runs_OnSupportedBackends(
+    public async Task Representative_Workflow_Runs_OnSupportedBackends(
         [ClassDataSource<AzuriteE2EBackendFixture>(Shared = SharedType.PerTestSession)] [ClassDataSource<AzureE2EBackendFixture>(Shared = SharedType.PerTestSession)] IE2EStorageBackend backend,
-        [MethodDataSource(typeof(RepresentativeScenarioCatalog), nameof(RepresentativeScenarioCatalog.All))] RepresentativeScenarioDefinition scenario,
+        [MethodDataSource(typeof(RepresentativeWorkflowCatalog), nameof(RepresentativeWorkflowCatalog.All))] RepresentativeWorkflowDefinition workflow,
         CancellationToken cancellationToken)
     {
         if (backend is AzureE2EBackendFixture && !AzureFixture.IsAvailable)
@@ -19,36 +19,15 @@ internal class RepresentativeArchiveRestoreTests
             return;
         }
 
-        if (ShouldSkipForAzureColdRestoreTimeout(backend, scenario))
-        {
-            Skip.Unless(false, $"Azure cold restore representative scenario is tracked by issue #65: {scenario.Name}");
-            return;
-        }
-
-        var result = await RepresentativeScenarioRunner.RunAsync(
+        var result = await RepresentativeWorkflowRunner.RunAsync(
             backend,
-            scenario,
-            SyntheticRepositoryProfile.Representative,
-            seed: 20260419,
-            dependencies: new RepresentativeScenarioRunnerDependencies
+            workflow,
+            dependencies: new RepresentativeWorkflowRunnerDependencies
             {
                 AssertRestoreTrees = true,
             },
             cancellationToken: cancellationToken);
 
-        if (scenario.BackendRequirement == ScenarioBackendRequirement.Any)
-            result.WasSkipped.ShouldBeFalse();
-    }
-
-    static bool ShouldSkipForAzureColdRestoreTimeout(IE2EStorageBackend backend, RepresentativeScenarioDefinition scenario)
-    {
-        if (backend is not AzureE2EBackendFixture)
-            return false;
-
-        return scenario == RepresentativeScenarioCatalog.RestoreLatestColdCache ||
-               scenario == RepresentativeScenarioCatalog.RestorePreviousColdCache ||
-               scenario == RepresentativeScenarioCatalog.RestoreLocalConflictNoOverwrite ||
-               scenario == RepresentativeScenarioCatalog.RestoreLocalConflictOverwrite ||
-               scenario == RepresentativeScenarioCatalog.ArchiveTierPlanning;
+        result.WasSkipped.ShouldBeFalse();
     }
 }
