+++ src/Arius.E2E.Tests/Workflows/RepresentativeWorkflowCatalogObjectIdentityTests.cs
@@ -1,0 +1,12 @@
+namespace Arius.E2E.Tests.Workflows;
+
+public class RepresentativeWorkflowCatalogObjectIdentityTests
+{
+    [Test]
+    public async Task Catalog_ExposesNamedWorkflowInstances_InAllCollection()
+    {
+        await Task.CompletedTask;
+
+        RepresentativeWorkflowCatalog.All.ShouldContain(RepresentativeWorkflowCatalog.Canonical);
+    }
+}
